Update index.html
Made design format changes as well as making a different layout. A new nav bar has been implemented to make it more responsive. Where in the old design there was no navigation bar on the header section. I changed the navigation bar so that when it is displayed on a larger screen, it flexes across the header and when it shifts to a smaller screen through the function of a hamburger a menu will pop up from the side.